functions/requirements.txt
ccxt
slackweb==1.0.5
pyti
PrettyTable 
google-cloud-datastore
